Walk/Run/Crouch
I added a :
1/.  Walk
2./  Crouch
3/.  Jump

and I updated the transitions to make the player walk if the float parameter "speed" surpass the value of 0.2. It will run if the bool parameter "run" is true and it return to walk or idle state depending on "speed" or "run" parameters.
And I made the player crouch from any state  if the bool parameter "crouch" became true and stay like that and won't return to idle state until "crouch" became false again.